# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MoneyExchangeTHGroup Admin API",
    "description": "RESTful API for the MoneyExchangeTHGroup Admin Panel. Manages currencies, exchange rates, members, bookings, wallet top-ups, articles, and admin profiles.\n\nAll endpoints (except `GET /api/currencies` and `GET /api/articles?public=true`) require authentication via Supabase session cookies.",
    "version": "1.0.0",
    "contact": {
      "name": "Nexmind IT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "description": "Authenticate with email and password. Sets Supabase session cookies on success.",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "admin@moneyexchangethgroup.com" },
                  "password": { "type": "string", "format": "password", "example": "••••••••" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": { "type": "object", "description": "Supabase user object" },
                        "session": { "type": "object", "description": "Supabase session object" }
                      }
                    },
                    "message": { "type": "string", "example": "Login successful" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/branches/{id}/margins": {
      "get": {
        "tags": ["Branches"],
        "summary": "Get branch margin overrides",
        "description": "Returns all currencies with their effective margins for this branch. Each entry indicates whether the margin is a branch-specific override or the global default.",
        "operationId": "getBranchMargins",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "silom" }
        ],
        "responses": {
          "200": {
            "description": "Branch margins",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "branch": { "$ref": "#/components/schemas/Branch" },
                        "margins": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/BranchMarginEntry" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "tags": ["Branches"],
        "summary": "Set all branch margin overrides",
        "description": "Bulk upsert margin overrides for a branch. All currencies in the array will be set/updated.",
        "operationId": "setBranchMargins",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "silom" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["margins"],
                "properties": {
                  "margins": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["currency_code", "buy_margin_percent", "sell_margin_percent"],
                      "properties": {
                        "currency_code": { "type": "string", "example": "USD" },
                        "buy_margin_percent": { "type": "number", "example": 2.0 },
                        "sell_margin_percent": { "type": "number", "example": 2.5 }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Margins updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/BranchCurrencyMargin" }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "patch": {
        "tags": ["Branches"],
        "summary": "Update single currency margin for branch",
        "description": "Upsert a margin override for a single currency at this branch.",
        "operationId": "updateBranchMargin",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "silom" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["currency_code"],
                "properties": {
                  "currency_code": { "type": "string", "example": "USD" },
                  "buy_margin_percent": { "type": "number", "example": 1.8 },
                  "sell_margin_percent": { "type": "number", "example": 2.2 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Margin updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/BranchCurrencyMargin" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/bookings": {
      "get": {
        "tags": ["Bookings"],
        "summary": "List all bookings",
        "operationId": "listBookings",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["pending_payment", "pending_review", "approved", "completed"] },
            "description": "Filter by booking status"
          }
        ],
        "responses": {
          "200": {
            "description": "Booking list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Booking" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Bookings"],
        "summary": "Create new booking",
        "operationId": "createBooking",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["member_id", "member_name", "currency_code", "amount", "rate", "total_thb"],
                "properties": {
                  "member_id": { "type": "string", "format": "uuid" },
                  "member_name": { "type": "string" },
                  "currency_code": { "type": "string", "example": "USD" },
                  "currency_flag": { "type": "string" },
                  "amount": { "type": "number", "example": 500 },
                  "rate": { "type": "number", "example": 34.50 },
                  "total_thb": { "type": "number", "example": 17250 },
                  "pickup_method": { "type": "string", "enum": ["branch", "wallet"], "default": "branch" },
                  "branch_name": { "type": "string", "nullable": true },
                  "pickup_date": { "type": "string", "format": "date-time", "nullable": true },
                  "status": { "type": "string", "enum": ["pending_payment", "pending_review", "approved", "completed"], "default": "pending_payment" },
                  "slip_url": { "type": "string", "format": "uri" },
                  "note": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Booking" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase session cookie set after login. Automatically managed by the browser."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Something went wrong" }
        }
      },
...
      "Branch": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "silom" },
          "name": { "type": "string", "example": "Silom" },
          "name_th": { "type": "string", "example": "สีลม" },
          "address": { "type": "string" },
          "hours": { "type": "string", "example": "09:00 – 20:00" },
          "status": { "type": "string", "enum": ["active", "inactive"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "BranchCurrencyMargin": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "branch_id": { "type": "string", "example": "silom" },
          "currency_code": { "type": "string", "example": "USD" },
          "buy_margin_percent": { "type": "number", "example": 2.0 },
          "sell_margin_percent": { "type": "number", "example": 2.5 },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "BranchMarginEntry": {
        "type": "object",
        "description": "Merged view of a currency's margin for a specific branch, showing both the effective margin and global defaults.",
        "properties": {
          "currency_code": { "type": "string", "example": "USD" },
          "currency_name": { "type": "string", "example": "US Dollar" },
          "currency_flag": { "type": "string", "example": "🇺🇸" },
          "buy_margin_percent": { "type": "number", "example": 2.0, "description": "Effective buy margin (branch override or global)" },
          "sell_margin_percent": { "type": "number", "example": 2.5, "description": "Effective sell margin (branch override or global)" },
          "is_override": { "type": "boolean", "description": "true if this branch has a custom margin for this currency" },
          "global_buy_margin": { "type": "number", "example": 1.5, "description": "Global default buy margin from currencies table" },
          "global_sell_margin": { "type": "number", "example": 1.5, "description": "Global default sell margin from currencies table" }
        }
      },
...
      "Booking": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "member_id": { "type": "string", "format": "uuid" },
          "member_name": { "type": "string" },
          "currency_code": { "type": "string", "example": "USD" },
          "currency_flag": { "type": "string" },
          "amount": { "type": "number", "example": 500 },
          "rate": { "type": "number", "example": 34.50 },
          "total_thb": { "type": "number", "example": 17250 },
          "pickup_method": { "type": "string", "enum": ["branch", "wallet"] },
          "branch_name": { "type": "string", "nullable": true },
          "pickup_date": { "type": "string", "format": "date-time", "nullable": true },
          "status": { "type": "string", "enum": ["pending_payment", "pending_review", "approved", "completed"] },
          "slip_url": { "type": "string", "format": "uri" },
          "note": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request — validation error or missing fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized — not logged in or session expired",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "Unauthorized" }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "Resource not found" }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "Internal server error" }
              }
            }
          }
        }
      }
    }
  }
}